// src/main/resources/angular/src/app/api/model/scheduleCityConf.ts
/**
 * Smou API
 * This is the Mobility as a Service API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: avalls@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { I18n } from './i18n';
import { VacationDay } from './vacationDay';


/**
 * Schedule of a parking area. Defined by days and the pairs of Start and End
 */
export interface ScheduleCityConf { 
    scheduleId: number;
    scheduleCode: string;
    /**
     * de Lun a Sáb 8-20
     */
    description?: string;
    description_i18n?: I18n;
    descriptionI18n?: I18n;
    /**
     * Data de la eliminació
     */
    deletedDate?: string;
    /**
     * Maximum length = 3
     */
    mondayStart?: Array<string>;
    /**
     * Maximum length = 3
     */
    mondayEnd?: Array<string>;
    tuesdayStart?: Array<string>;
    tuesdayEnd?: Array<string>;
    wednesdayStart?: Array<string>;
    wednesdayEnd?: Array<string>;
    thursdayStart?: Array<string>;
    thursdayEnd?: Array<string>;
    fridayStart?: Array<string>;
    fridayEnd?: Array<string>;
    saturdayStart?: Array<string>;
    saturdayEnd?: Array<string>;
    sundayStart?: Array<string>;
    sundayEnd?: Array<string>;
    inUse?: boolean;
    vacationDay?: Array<VacationDay>;

}