// src/main/resources/angular/src/app/api/model/segmentAll.ts
/**
 * Dum API
 * This is the DUM API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: rbarbado@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ConfigurationAll } from './configurationAll';


/**
 * Entire information of a segment
 */
export interface SegmentAll { 
    /**
     * Segment ID
     */
    segmentId?: number;
    /**
     * Description
     */
    description?: string;
    /**
     * Geometry
     */
    geometry?: string;
    /**
     * City ID (NOT REQUIRED)
     */
    cityId?: number;
    /**
     * Places
     */
    places?: number;
    /**
     * Name
     */
    name?: string;
    maxTime: number;
    maxTimeInfinity: boolean;
    scheduleId: number;
    zoneTypeId: number;
    segmentCode:string;
}