// src/main/resources/angular/src/app/api/model/segmentEdit.ts
/**
 * Dum API
 * This is the DUM API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: rbarbado@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Segment edit model
 */
export interface SegmentEdit { 
    /**
     * Description
     */
    description?: string;
    /**
     * Geometry
     */
    geometry?: string;
    /**
     * City ID
     */
    cityId?: number;
    /**
     * Configuration
     */
    configurationId?: number;
    /**
     * Places
     */
    places?: number;
    /**
     * Name
     */
    name?: string;
    maxTime: number;
    maxTimeInfinity: boolean;
    scheduleId: number;
    zoneTypeId: number;
    codeSegment: string;
}