// src/main/resources/angular/src/app/api/model/zoneType.ts
/**
 * Smou API
 * This is the Mobility as a Service API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: avalls@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { I18n } from './i18n';


export interface ZoneType {
    /**
     * 1→ Blava, 2→ ?, 3→ Verda
     */
    zoneTypeId: number;
    name: string;
    rgb: string;
    description?: I18n;
}